% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CardinalityGame.R
\name{cardinalityGameValue}
\alias{cardinalityGameValue}
\title{Compute value of a coalition for a cardinality game}
\usage{
cardinalityGameValue(S)
}
\arguments{
\item{S}{numeric vector with coalition of players}
}
\value{
The return value is the cardinality, i.e. the number of elements, of coalition S
}
\description{
\strong{Coalition value for a cardinality game:} \cr
For further information see \link{cardinalityGame}
}
\examples{
library(CoopGame)
S=c(1,2,4,5)
cardinalityGameValue(S)
}
\author{
Johannes Anwander \email{anwander.johannes@gmail.com}

Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
