% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CardinalityGame.R
\name{cardinalityGameVector}
\alias{cardinalityGameVector}
\title{Compute game vector for a cardinality game}
\usage{
cardinalityGameVector(n)
}
\arguments{
\item{n}{represents the number of players}
}
\value{
Game vector for the cardinality game
}
\description{
\strong{Game vector for a cardinality game:} \cr
For further information see \link{cardinalityGame}
}
\examples{
library(CoopGame)
cardinalityGameVector(n=3)

\donttest{
library(CoopGame)
(v <- cardinalityGameVector(n=4))
#[1] 1 1 1 1 2 2 2 2 2 2 3 3 3 3 4
}

}
\author{
Johannes Anwander \email{anwander.johannes@gmail.com}

Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
