% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivideTheDollarGame.R
\name{divideTheDollarGameVector}
\alias{divideTheDollarGameVector}
\title{Compute game vector for a divide-the-dollar game}
\usage{
divideTheDollarGameVector(n)
}
\arguments{
\item{n}{represents the number of players}
}
\value{
Game vector for the specified divide-the-dollar game
}
\description{
\strong{Game vector for a divide-the-dollar game:} \cr
For further information see \link{divideTheDollarGame}
}
\examples{
library(CoopGame) 
divideTheDollarGameVector(n=3)

\donttest{
library(CoopGame)
(v <- divideTheDollarGameVector(n=4))
#Output: 
# [1] 0 0 0 0 1 1 1 1 1 1 1 1 1 1 1
}

}
\references{
Jackson M.O. (2008)  \emph{Social and Economic Networks}, Princeton University Press, p. 413
}
\author{
Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}

Johannes Anwander \email{anwander.johannes@gmail.com}
}
