% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGapFunctionCoefficients.R
\name{getGapFunctionCoefficients}
\alias{getGapFunctionCoefficients}
\title{Compute gap function coefficients}
\usage{
getGapFunctionCoefficients(v)
}
\arguments{
\item{v}{Numeric vector of length 2^n - 1 representing the values of 
the coalitions of a TU game with n players}
}
\value{
numeric vector containing the gap function coefficients for every coalition
}
\description{
getGapFunctionCoefficients computes the gap function coefficients for a 
specified TU game
}
\examples{
library(CoopGame)
getGapFunctionCoefficients(c(0,0,0,60,48,30,72))

}
\references{
Driessen T. (1998) \emph{Cooperative Games, Solutions and Applications}, Springer, p. 57
}
\author{
Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
