% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameVector.R
\name{getNumberOfPlayers}
\alias{getNumberOfPlayers}
\title{Get number of players}
\usage{
getNumberOfPlayers(v)
}
\arguments{
\item{v}{Numeric vector of length 2^n - 1 representing the values of 
the coalitions of a TU game with n players}
}
\value{
Number of players in the game (specified by game vector v)
}
\description{
Gets the number of players from a game vector
}
\examples{
library(CoopGame)
maschlerGame=c(0,0,0,60,60,60,72)
getNumberOfPlayers(maschlerGame)

}
\author{
Michael Maerz

Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
