% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPerCapitaExcessCoefficients.R
\name{getPerCapitaExcessCoefficients}
\alias{getPerCapitaExcessCoefficients}
\title{Compute per capita excess coefficients}
\usage{
getPerCapitaExcessCoefficients(v, x)
}
\arguments{
\item{v}{Numeric vector of length 2^n - 1 representing the values of 
the coalitions of a TU game with n players}

\item{x}{numeric vector containing allocations for each player}
}
\value{
numeric vector containing the per capita excess coefficients for every coalition
}
\description{
getPerCapitaExcessCoefficients computes the 
per capita excess coefficients for a 
specified TU game and an allocation x
}
\examples{
library(CoopGame)
getPerCapitaExcessCoefficients(c(0,0,0,60,48,30,72), c(24,24,24))

}
\author{
Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
