% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitMatrix.R
\name{getPlayersFromBitVector}
\alias{getPlayersFromBitVector}
\title{Extract players from bit vector}
\usage{
getPlayersFromBitVector(bitVector)
}
\arguments{
\item{bitVector}{represents the binary vector}
}
\value{
playerVector contains the numbers of the players involved in the coalition
}
\description{
getPlayersFromBitVector determines players involved in a coalition
from a binary vector.
}
\examples{
library(CoopGame)
myBitVector <-c(1,0,1,0)
(players<-getPlayersFromBitVector(myBitVector))

}
\author{
Johannes Anwander \email{anwander.johannes@gmail.com}

Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
