% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloveGame.R
\name{gloveGameVector}
\alias{gloveGameVector}
\title{Compute game vector for glove game}
\usage{
gloveGameVector(n, L, R)
}
\arguments{
\item{n}{represents the number of players}

\item{L}{numeric vector of players owning one left-hand glove each}

\item{R}{numeric vector of players owning one right-hand glove each}
}
\value{
Game vector of the specified glove game
}
\description{
\strong{Game vector for glove game: }\cr
For further information see \link{gloveGame}
}
\examples{
library(CoopGame)
gloveGameVector(3, L=c(1,2), R=c(3))

}
\references{
Peters H. (2015)  \emph{Game Theory: A Multi-Leveled Approach}, 2nd Edition, Springer, pp. 155--156
}
\author{
Johannes Anwander \email{anwander.johannes@gmail.com}

Jochen Staudacher \email{jochen.staudacher@hs-kempten.de}
}
